refactored so that all msbuild related code is in MSBuildAPIUtility
