+++ test/NuGet.Core.Tests/NuGet.CommandLine.Xplat.Tests/XPlatWhyUnitTests.cs
@@ -2,27 +2,17 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using Microsoft.Build.Locator;
 using NuGet.CommandLine.XPlat;
-using NuGet.CommandLine.Xplat.Tests.Utility;
-using NuGet.Frameworks;
 using NuGet.ProjectModel;
 using NuGet.Test.Utility;
 using Test.Utility;
 using Xunit;
-using System.Globalization;
 
 namespace NuGet.CommandLine.Xplat.Tests
 {
     public class XPlatWhyUnitTests
     {
-        static XPlatWhyUnitTests()
-        {
-            MSBuildLocator.RegisterDefaults();
-        }
-
         [Fact]
         public void WhyCommand_DependencyGraphFinder_MultipleDependencyPathsForTargetPackage_AllPathsFound()
         {
@@ -214,220 +204,6 @@ namespace NuGet.CommandLine.Xplat.Tests
 
             // transitive dependency from a top-level project reference is found, with the correct resolved version
             Assert.Contains(dependencyGraphs["net472"].First(dep => dep.Id == "DotnetNuGetWhyPackage").Children, dep => (dep.Id == "System.Text.Json") && (dep.Version == "8.0.0"));
-        }
-
-        [Fact]
-        public void WhyCommand_GetListOfProjectPaths_WithProjectFile_ReturnsCorrectPaths()
-        {
-            // Arrange
-            var pathContext = new SimpleTestPathContext();
-            var testLogger = new XPlatTestLogger();
-
-            var net8 = NuGetFramework.Parse("net8.0");
-
-            var projectA = SimpleTestProjectContext.CreateNETCore(
-                "a",
-                pathContext.SolutionRoot,
-                net8);
-
-            projectA.Save();
-
-            var whyCommandArgs = new WhyCommandArgs(
-                path: projectA.ProjectPath,
-                package: "package",
-                frameworks: new List<string>(),
-                logger: testLogger);
-
-            // Act
-            var projectList = whyCommandArgs.GetListOfProjectPaths();
-
-            // Assert
-            Assert.NotNull(projectList);
-            Assert.Equal(projectList.Count(), 1);
-
-            Assert.Contains(projectA.ProjectPath, projectList);
-        }
-
-        [Fact]
-        public void WhyCommand_GetListOfProjectPaths_WithProjectDirectory_ReturnsCorrectPaths()
-        {
-            // Arrange
-            var pathContext = new SimpleTestPathContext();
-            var testLogger = new XPlatTestLogger();
-
-            var net8 = NuGetFramework.Parse("net8.0");
-
-            var projectA = SimpleTestProjectContext.CreateNETCore(
-                "a",
-                pathContext.SolutionRoot,
-                net8);
-
-            projectA.Save();
-
-            var whyCommandArgs = new WhyCommandArgs(
-                path: Path.GetDirectoryName(projectA.ProjectPath),
-                package: "package",
-                frameworks: new List<string>(),
-                logger: testLogger);
-
-            // Act
-            var projectList = whyCommandArgs.GetListOfProjectPaths();
-
-            // Assert
-            Assert.NotNull(projectList);
-            Assert.Equal(projectList.Count(), 1);
-
-            Assert.Contains(projectA.ProjectPath, projectList);
-        }
-
-        [Fact]
-        public void WhyCommand_GetListOfProjectPaths_WithSolutionFile_ReturnsCorrectPaths()
-        {
-            // Arrange
-            var pathContext = new SimpleTestPathContext();
-            var testLogger = new XPlatTestLogger();
-
-            var solution = new SimpleTestSolutionContext(pathContext.SolutionRoot);
-
-            var net8 = NuGetFramework.Parse("net8.0");
-
-            var projectA = SimpleTestProjectContext.CreateNETCore(
-                "a",
-                pathContext.SolutionRoot,
-                net8);
-            var projectB = SimpleTestProjectContext.CreateNETCore(
-                "b",
-                pathContext.SolutionRoot,
-                net8);
-
-            solution.Projects.Add(projectA);
-            solution.Projects.Add(projectB);
-            solution.Create(pathContext.SolutionRoot);
-
-            var whyCommandArgs = new WhyCommandArgs(
-                path: solution.SolutionPath,
-                package: "package",
-                frameworks: new List<string>(),
-                logger: testLogger);
-
-            // Act
-            var projectList = whyCommandArgs.GetListOfProjectPaths();
-
-            // Assert
-            Assert.NotNull(projectList);
-            Assert.Equal(projectList.Count(), 2);
-
-            Assert.Contains(projectA.ProjectPath, projectList);
-            Assert.Contains(projectB.ProjectPath, projectList);
-        }
-
-        [Fact]
-        public void WhyCommand_GetListOfProjectPaths_WithSolutionDirectory_ReturnsCorrectPaths()
-        {
-            // Arrange
-            var pathContext = new SimpleTestPathContext();
-            var testLogger = new XPlatTestLogger();
-
-            var solution = new SimpleTestSolutionContext(pathContext.SolutionRoot);
-
-            var net8 = NuGetFramework.Parse("net8.0");
-
-            var projectA = SimpleTestProjectContext.CreateNETCore(
-                "a",
-                pathContext.SolutionRoot,
-                net8);
-            var projectB = SimpleTestProjectContext.CreateNETCore(
-                "b",
-                pathContext.SolutionRoot,
-                net8);
-
-            solution.Projects.Add(projectA);
-            solution.Projects.Add(projectB);
-            solution.Create(pathContext.SolutionRoot);
-
-            var whyCommandArgs = new WhyCommandArgs(
-                path: pathContext.SolutionRoot,
-                package: "package",
-                frameworks: new List<string>(),
-                logger: testLogger);
-
-            // Act
-            var projectList = whyCommandArgs.GetListOfProjectPaths();
-
-            // Assert
-            Assert.NotNull(projectList);
-            Assert.Equal(projectList.Count(), 2);
-
-            Assert.Contains(projectA.ProjectPath, projectList);
-            Assert.Contains(projectB.ProjectPath, projectList);
-        }
-
-        [Theory]
-        [InlineData("X.sln", "Y.sln")]
-        [InlineData("A.csproj", "B.csproj")]
-        [InlineData("X.sln", "A.csproj")]
-        public void WhyCommand_GetListOfProjectPaths_WithDirectoryWithMultipleSolutionsOrProjects_ReturnsNull(params string[] directoryFiles)
-        {
-            // Arrange
-            var pathContext = new SimpleTestPathContext();
-            var testLogger = new XPlatTestLogger();
-
-            var solution = new SimpleTestSolutionContext(pathContext.SolutionRoot);
-
-            foreach (var filename in directoryFiles)
-            {
-                var filePath = Path.Combine(pathContext.SolutionRoot, filename);
-                File.Create(filePath);
-            }
-
-            var whyCommandArgs = new WhyCommandArgs(
-                path: pathContext.SolutionRoot,
-                package: "package",
-                frameworks: new List<string>(),
-                logger: testLogger);
-
-            var expectedErrorMessage = string.Format(CultureInfo.CurrentCulture,
-                $"Unable to run 'dotnet nuget why'. The directory '{pathContext.SolutionRoot}' contains multiple project or solution files. Please specify which project or solution file to use.");
-
-            // Act
-            var projectList = whyCommandArgs.GetListOfProjectPaths();
-
-            // Assert
-            var errors = testLogger.ShowErrors();
-
-            Assert.Null(projectList);
-            Assert.Contains(expectedErrorMessage, errors);
-        }
-
-        [Fact]
-        public void WhyCommand_GetListOfProjectPaths_WithDirectoryWithZeroSolutionsOrProjects_ReturnsNull()
-        {
-            // Arrange
-            var pathContext = new SimpleTestPathContext();
-            var testLogger = new XPlatTestLogger();
-
-            var solution = new SimpleTestSolutionContext(pathContext.SolutionRoot);
-
-            var randomFilePath = Path.Combine(pathContext.SolutionRoot, "x.text");
-            File.Create(randomFilePath);
-
-            var whyCommandArgs = new WhyCommandArgs(
-                path: pathContext.SolutionRoot,
-                package: "package",
-                frameworks: new List<string>(),
-                logger: testLogger);
-
-            var expectedErrorMessage = string.Format(CultureInfo.CurrentCulture,
-                $"Unable to run 'dotnet nuget why'. The directory '{pathContext.SolutionRoot}' does not contain any project or solution files.");
-
-            // Act
-            var projectList = whyCommandArgs.GetListOfProjectPaths();
-
-            // Assert
-            var errors = testLogger.ShowErrors();
-
-            Assert.Null(projectList);
-            Assert.Contains(expectedErrorMessage, errors);
         }
 
         private static void ConvertRelevantWindowsPathsToUnix(LockFile assetsFile)
